+++ remove_database_user.sh
@@ -32,24 +32,24 @@ done
 #if version, display version and exit 
 [ -n "$opt_version" ] && echo "Version $(basename $0) $VERSION" && exit 0
 #if no client or no email, then exit
-[ -z "$opt_domain" ] && echo "ERROR : Domain name is missing" && exit 1
-[ -z "$opt_user" ] && echo "ERROR : User name is missing" && exit 1
+[ -z "$opt_domain" ] && error "Domain name is missing"
+[ -z "$opt_user" ] && error "User name is missing"
 
 #argument vs system ckeckings :
-[ -z "`query "select name from domains where name='$opt_domain_val';"`" ] && echo "ERROR : Domain $opt_domail_val is unknown" && exit 1
-[ -z "`query "select domain from database_domains where domain='$opt_domain_val';"`" ] && echo "ERROR : Service database for domain $opt_domain_val is disabled" && exit 1
-[ -z "`query "select name from database_users where domain='$opt_domain_val' and name='$opt_user_val'"`" ] && echo "ERROR : User $opt_user_val not defined for domain $opt_domain_val" && exit 1
+[ -z "`query "select name from domains where name='$opt_domain_val';"`" ] && error "Domain $opt_domail_val is unknown"
+[ -z "`query "select domain from database_domains where domain='$opt_domain_val';"`" ] && error "Service database for domain $opt_domain_val is disabled"
+[ -z "`query "select name from database_users where domain='$opt_domain_val' and name='$opt_user_val'"`" ] && error "User $opt_user_val not defined for domain $opt_domain_val"
 
 #verif
 opt_user_val=`query "select name from database_users where domain='$opt_domain_val' and name='$opt_user_val';"`
 
 #deleting user entry
-query "delete from database_users where name='$opt_user_val';" || exit 1
+query "delete from database_users where name='$opt_user_val';" error "Client integrity at risk; aborting"
 
 #upgrading system level
 mysql -N -u$DATABASE_ADMIN_USER -p$DATABASE_ADMIN_PASS -e "use mysql ; drop user '$opt_user_val'@'localhost' ;" && exit 0
 
 #otherwise, something went wrong.
-echo "ERROR : something unexpected appened" && exit 1
+error "something unexpected appened"
 #peut etre effacer içi l'enregistrement en bdd ??
